big update

+++ src/features/admin/categories/server/routers.ts
@@ -1,3 +1,4 @@
+import { generateCategorySlug } from "@/lib/generate-category-slug";
 import { prisma } from "@/lib/prisma";
 import { base, admin } from "@/orpc/init";
 import { ORPCError } from "@orpc/client";
@@ -10,14 +11,39 @@ export const categoriesRouter = base.router({
                 name: z.string().min(1),
                 billboardId: z.string().min(1),
                 storeId: z.string().min(1),
+                subcategories: z.array(z.string().min(1)),
             }),
         )
         .handler(async ({ input }) => {
+            const parenSlug = generateCategorySlug(input.name);
+            const existing = await prisma.category.findUnique({
+                where: {
+                    storeId_slug: {
+                        storeId: input.storeId,
+                        slug: parenSlug,
+                    },
+                },
+            });
+
+            if (existing) {
+                throw new ORPCError("CONFLICT");
+            }
+
             const categoryCreated = await prisma.category.create({
                 data: {
                     name: input.name,
                     billboardId: input.billboardId,
                     storeId: input.storeId,
+                    slug: parenSlug,
+                    children: {
+                        createMany: {
+                            data: input.subcategories.map((sub) => ({
+                                name: sub,
+                                storeId: input.storeId,
+                                slug: generateCategorySlug(input.name, sub),
+                            })),
+                        },
+                    },
                 },
             });
 
@@ -28,7 +54,8 @@ export const categoriesRouter = base.router({
             z.object({
                 id: z.string().min(1),
                 name: z.string().min(1),
-                billboardId: z.string().min(1),
+                billboardId: z.string().min(1).nullable(),
+                subcategories: z.array(z.string().min(1)),
                 storeId: z.string().min(1),
             }),
         )
@@ -38,20 +65,72 @@ export const categoriesRouter = base.router({
                     id: input.id,
                     storeId: input.storeId,
                 },
+                include: {
+                    children: {
+                        include: {
+                            _count: {
+                                select: { products: true },
+                            },
+                        },
+                    },
+                },
             });
 
             if (!category) {
                 throw new ORPCError("NOT_FOUND");
             }
 
-            const categoryUpdated = await prisma.category.update({
-                where: {
-                    id: input.id,
-                },
-                data: {
-                    name: input.name,
-                    billboardId: input.billboardId,
-                },
+            const categoryUpdated = await prisma.$transaction(async (tx) => {
+                const currentChildren = category.children;
+                const currentChildrenSlugs = currentChildren.map((child) => child.slug);
+
+                const subcategoriesInputFormat = input.subcategories.map((sub) => ({
+                    name: sub,
+                    slug: generateCategorySlug(category.name, sub.trim().toLowerCase()),
+                }));
+
+                const subcategoriesInputFormatSlugs = input.subcategories.map((sub) => generateCategorySlug(category.name, sub.trim().toLowerCase()));
+
+                const subcategoriesToDelete = currentChildren
+                    .filter((child) => !subcategoriesInputFormatSlugs.includes(child.slug) && child._count.products === 0)
+                    .map((child) => child.slug);
+
+                const subcategoriesToCreate = subcategoriesInputFormat.filter((sub) => !currentChildrenSlugs.includes(sub.slug));
+
+                if (subcategoriesToDelete.length > 0) {
+                    await tx.category.deleteMany({
+                        where: {
+                            parentId: input.id,
+                            slug: {
+                                in: subcategoriesToDelete,
+                            },
+                        },
+                    });
+                }
+
+                const parenSlug = generateCategorySlug(input.name);
+
+                return await tx.category.update({
+                    where: {
+                        id: input.id,
+                    },
+                    data: {
+                        name: input.name,
+                        billboardId: input.billboardId,
+                        slug: parenSlug,
+                        ...(subcategoriesToCreate.length > 0 && {
+                            children: {
+                                createMany: {
+                                    data: subcategoriesToCreate.map((sub) => ({
+                                        name: sub.name,
+                                        storeId: input.storeId,
+                                        slug: sub.slug,
+                                    })),
+                                },
+                            },
+                        }),
+                    },
+                });
             });
 
             return categoryUpdated;
@@ -69,10 +148,29 @@ export const categoriesRouter = base.router({
                     id: input.id,
                     storeId: input.storeId,
                 },
+                include: {
+                    children: {
+                        include: {
+                            _count: {
+                                select: { products: true },
+                            },
+                        },
+                    },
+                },
             });
 
             if (!category) {
                 throw new ORPCError("NOT_FOUND");
+            }
+
+            const subsWithProducts = category.children.filter((sub) => sub._count.products > 0);
+
+            if (subsWithProducts.length > 0) {
+                const invalidNames = subsWithProducts.map((sub) => sub.name).join(", ");
+
+                throw new ORPCError("BAD_REQUEST", {
+                    message: `Can't delete subcategories: [${invalidNames}] because they contain the product.`,
+                });
             }
 
             const categoryDeleted = await prisma.category.delete({
@@ -157,17 +255,27 @@ export const categoriesRouter = base.router({
                     id: input.id,
                     storeId: input.storeId,
                 },
+                include: {
+                    children: {
+                        include: {
+                            _count: {
+                                select: { products: true },
+                            },
+                        },
+                    },
+                },
             });
 
             return category;
         }),
-    getMany: admin.input(z.object({ storeId: z.string().min(1) })).handler(async ({ input }) => {
+    getManyParent: admin.input(z.object({ storeId: z.string().min(1) })).handler(async ({ input }) => {
         const categories = await prisma.category.findMany({
             where: {
                 storeId: input.storeId,
+                parentId: null,
             },
             include: {
-                billboard: true,
+                children: true,
             },
             orderBy: {
                 createdAt: "desc",
@@ -176,22 +284,22 @@ export const categoriesRouter = base.router({
 
         return categories;
     }),
-    getManyWithPromotion: admin
-        .input(z.object({ storeId: z.string().min(1) }))
-        .handler(async ({ input }) => {
-            const categories = await prisma.category.findMany({
-                where: {
-                    storeId: input.storeId,
-                },
-                include: {
-                    billboard: true,
-                    promotions: true,
-                },
-                orderBy: {
-                    createdAt: "desc",
-                },
-            });
+    getManyWithPromotion: admin.input(z.object({ storeId: z.string().min(1) })).handler(async ({ input }) => {
+        const categories = await prisma.category.findMany({
+            where: {
+                storeId: input.storeId,
+                parent: null,
+            },
+            include: {
+                billboard: true,
+                promotions: true,
+                children: true,
+            },
+            orderBy: {
+                createdAt: "desc",
+            },
+        });
 
-            return categories;
-        }),
+        return categories;
+    }),
 });
